requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
bcrypt==4.0.1
python-jose==3.3.0
